requirements.txt.txt
streamlit
opencv-python-headless
mediapipe
numpy
pillow